refactor: UserPlan 도메인 분리
