+++ src/main/java/travel/travelapplication/dto/place/LikedPlaceList.java
@@ -1,4 +1,4 @@
-package travel.travelapplication.dto.userplan;
+package travel.travelapplication.dto.place;
 
 import lombok.Getter;
 import lombok.Setter;
