+++ src/main/java/travel/travelapplication/place/presentation/RecommendationController.java
@@ -2,18 +2,11 @@ package travel.travelapplication.place.presentation;
 
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import reactor.core.publisher.Mono;
 import travel.travelapplication.place.application.RecommendationService;
 import travel.travelapplication.place.domain.Recommendation;
-import travel.travelapplication.user.application.UserPlanService;
-import travel.travelapplication.user.domain.User;
-import travel.travelapplication.user.domain.UserPlan;
 
 @RestController
 public class RecommendationController {
