+++ src/main/java/travel/travelapplication/userplan/dto/UserPlanInfoRequest.java
@@ -1,0 +1,14 @@
+package travel.travelapplication.userplan.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+public record UserPlanInfoRequest (
+    String email,
+    String city,
+    String district,
+    long period
+){
+
+}
