+++ src/main/java/travel/travelapplication/userplan/repository/UserPlanRepository.java
@@ -1,10 +1,10 @@
-package travel.travelapplication.user.repository;
+package travel.travelapplication.userplan.repository;
 
 import org.bson.types.ObjectId;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import travel.travelapplication.user.domain.UserPlan;
+import travel.travelapplication.userplan.domain.UserPlan;
 
 public interface UserPlanRepository extends MongoRepository<UserPlan, ObjectId> {
